+++ identity/webapp/views/pages/delete-requested/index.tsx
@@ -9,7 +9,7 @@ import {
   Container,
   SectionHeading,
   Wrapper,
-} from '@weco/identity/views/components/styled/ALayouts';
+} from '@weco/identity/views/components/styled/Layouts';
 import IdentityPageLayout from '@weco/identity/views/layouts/IdentityPageLayout';
 
 const DeleteRequestedPage: NextPage = () => {
