+++ identity/webapp/views/pages/index.ChangeEmail.tsx
@@ -17,12 +17,12 @@ import { validEmailPattern } from '@weco/identity/utils/validation-patterns';
 import { ChangeDetailsModalContentProps } from '@weco/identity/views/components/ChangeDetailsModal';
 import Loading from '@weco/identity/views/components/Loading';
 import { PasswordInput } from '@weco/identity/views/components/PasswordInput';
-import { StatusAlert } from '@weco/identity/views/components/styled/AAlert';
-import { FieldMargin } from '@weco/identity/views/components/styled/AForms';
+import { StatusAlert } from '@weco/identity/views/components/styled/Alert';
+import { FieldMargin } from '@weco/identity/views/components/styled/Forms';
 import {
   ModalContainer,
   ModalTitle,
-} from '@weco/identity/views/components/styled/Amodal';
+} from '@weco/identity/views/components/styled/Modal';
 
 type ChangeEmailInputs = {
   email: string;
