+++ identity/webapp/views/pages/index.ChangePassword.tsx
@@ -15,12 +15,12 @@ import { ChangeDetailsModalContentProps } from '@weco/identity/views/components/
 import Loading from '@weco/identity/views/components/Loading';
 import { PasswordInput } from '@weco/identity/views/components/PasswordInput';
 import PasswordRules from '@weco/identity/views/components/PasswordRules';
-import { StatusAlert } from '@weco/identity/views/components/styled/AAlert';
-import { FieldMargin } from '@weco/identity/views/components/styled/AForms';
+import { StatusAlert } from '@weco/identity/views/components/styled/Alert';
+import { FieldMargin } from '@weco/identity/views/components/styled/Forms';
 import {
   ModalContainer,
   ModalTitle,
-} from '@weco/identity/views/components/styled/AModal';
+} from '@weco/identity/views/components/styled/Modal';
 
 type ChangePasswordInputs = {
   password: string;
