+++ identity/webapp/views/pages/index.tsx
@@ -30,12 +30,12 @@ import Loading from '@weco/identity/views/components/Loading';
 import {
   StatusAlert,
   StatusAlertProps,
-} from '@weco/identity/views/components/styled/AAlert';
+} from '@weco/identity/views/components/styled/Alert';
 import {
   Container,
   SectionHeading,
   Wrapper,
-} from '@weco/identity/views/components/styled/ALayouts';
+} from '@weco/identity/views/components/styled/Layouts';
 import IdentityPageLayout from '@weco/identity/views/layouts/IdentityPageLayout';
 
 import ChangeEmail from './index.ChangeEmail';
